+++ src/blob/BlobRequestListenerFactory.ts
@@ -31,6 +31,8 @@ import IBlobMetadataStore from "./persistence/IBlobMetadataStore";
 import { DEFAULT_CONTEXT_PATH } from "./utils/constants";
 
 import morgan = require("morgan");
+import { OAuthLevel } from "../common/models";
+import IAuthenticator from "./authentication/IAuthenticator";
 
 /**
  * Default RequestListenerFactory based on express framework.
@@ -50,7 +52,8 @@ export default class BlobRequestListenerFactory
     private readonly accountDataStore: IAccountDataStore,
     private readonly enableAccessLog: boolean,
     private readonly accessLogWriteStream?: NodeJS.WritableStream,
-    private readonly loose?: boolean
+    private readonly loose?: boolean,
+    private readonly oauth?: OAuthLevel
   ) {}
 
   public createRequestListener(): RequestListener {
@@ -134,22 +137,29 @@ export default class BlobRequestListenerFactory
     const authenticationMiddlewareFactory = new AuthenticationMiddlewareFactory(
       logger
     );
+    const authenticators: IAuthenticator[] = [
+      new PublicAccessAuthenticator(this.metadataStore, logger),
+      new BlobSharedKeyAuthenticator(this.accountDataStore, logger),
+      new AccountSASAuthenticator(
+        this.accountDataStore,
+        this.metadataStore,
+        logger
+      ),
+      new BlobSASAuthenticator(
+        this.accountDataStore,
+        this.metadataStore,
+        logger
+      )
+    ];
+    if (this.oauth !== undefined) {
+      authenticators.push(
+        new BlobTokenAuthenticator(this.accountDataStore, this.oauth, logger)
+      );
+    }
     app.use(
-      authenticationMiddlewareFactory.createAuthenticationMiddleware([
-        new PublicAccessAuthenticator(this.metadataStore, logger),
-        new BlobSharedKeyAuthenticator(this.accountDataStore, logger),
-        new AccountSASAuthenticator(
-          this.accountDataStore,
-          this.metadataStore,
-          logger
-        ),
-        new BlobSASAuthenticator(
-          this.accountDataStore,
-          this.metadataStore,
-          logger
-        ),
-        new BlobTokenAuthenticator(this.accountDataStore, logger)
-      ])
+      authenticationMiddlewareFactory.createAuthenticationMiddleware(
+        authenticators
+      )
     );
 
     // Generated, will do basic validation defined in swagger