+++ src/queue/QueueConfiguration.ts
@@ -36,7 +36,8 @@ export default class QueueConfiguration extends ConfigurationBase {
     loose: boolean = false,
     cert: string = "",
     key: string = "",
-    pwd: string = ""
+    pwd: string = "",
+    oauth?: string
   ) {
     super(
       host,
@@ -48,7 +49,8 @@ export default class QueueConfiguration extends ConfigurationBase {
       loose,
       cert,
       key,
-      pwd
+      pwd,
+      oauth
     );
   }
 }